requirements.txt
transformers==4.46.3
torch==2.5.1
torch-geometric==2.6.1
numpy==2.1.3
biopython==1.84
requests==2.32.3
matplotlib==3.9.2
# circuitbreaker==2.0.0
# httpx==0.28.0
# retry-async==0.1.4